src/mumble: replace target_link_libraries() for Opus and CELT
target_link_libraries() automatically adds the library's include directories to the specified target and marks it as dependency.

Both steps are fine, however there's a third we don't need: linking to the library, because we load Opus and CELT at runtime.

To make it clear that we don't link to the libraries, this commit replaces target_link_libraries() with add_dependencies() and target_include_directories().